feat: Improve build pipeline (#78)
This pull request enhances the `.github/workflows/update-flake.yaml`
file by adding new features and improving the build and validation
process. The key changes include allowing manual triggering, adding a
matrix strategy to run jobs on multiple operating systems, and
implementing steps to handle build and check failures by creating issues
automatically.

Enhancements to build and validation process:

*
[`.github/workflows/update-flake.yaml`](diffhunk://#diff-4a139fc25db101424a99bb7668e55ce5a6a6f32aa9a3ed5838b638a3d14e4a4cR1-R68):
Added `workflow_dispatch` to allow manual triggering of the workflow.
*
[`.github/workflows/update-flake.yaml`](diffhunk://#diff-4a139fc25db101424a99bb7668e55ce5a6a6f32aa9a3ed5838b638a3d14e4a4cR1-R68):
Introduced a matrix strategy to run jobs on both `ubuntu-latest` and
`macos-latest`, and included `fail-fast: false` to continue with other
jobs even if one fails.
*
[`.github/workflows/update-flake.yaml`](diffhunk://#diff-4a139fc25db101424a99bb7668e55ce5a6a6f32aa9a3ed5838b638a3d14e4a4cR1-R68):
Added a step to build and test the configuration, capturing the output
and setting the build status in the environment variables.
*
[`.github/workflows/update-flake.yaml`](diffhunk://#diff-4a139fc25db101424a99bb7668e55ce5a6a6f32aa9a3ed5838b638a3d14e4a4cR1-R68):
Implemented a step to create an issue on build failure, using
`actions/github-script@v7` to post the build output in the issue
description.
*
[`.github/workflows/update-flake.yaml`](diffhunk://#diff-4a139fc25db101424a99bb7668e55ce5a6a6f32aa9a3ed5838b638a3d14e4a4cR77-R120):
Added steps to run checks and create an issue on check failure, similar
to the build failure handling.